Added increasing / decreasing speed. It increases when you press the forward button, decreases if you press back, it increases faster if you are still going backwards and decreases faster if you are still going forwards, if no keys are pressed, it accelerates to 0 if you were going backwards and decelerates to 0 if you were going forwards. The turning speed is also variable based on the current speed of the car, so if you are going slow, you can only turn slow.
Made a max speed and a current speed variable (used for checking if at maxSpeed or at -maxSpeed).

Also added a feature to do some mad skiddies (smoke to be included).

Changed the car selection time form 5 seconds to 3.

Changed the frequency of changing controls from every 30 seconds, to every minute (will test it out).

Fixed the length of the Raycast from the miniaturising of the robot and cars in the Garage.

Fancied up the Garage a bit more, was kinda bland...